add: snippet repository change: refactor snippets code
