+++ lang/en/theme_boost_union.php
@@ -1376,7 +1376,7 @@ $string['snippetssnippets'] = 'SCSS snippets';
 $string['snippetssource'] = 'Source';
 $string['snippetssourcetheme_boost_union'] = 'Boost Union built-in';
 $string['snippetssourceuploaded'] = 'Upload';
-$string['snippetssourcecommunity_repository'] = 'Community Repository';
+$string['snippetssourcecommunity'] = 'Community Repository';
 $string['snippetstestedon'] = 'Tested on';
 $string['snippetstitle'] = 'Title';
 $string['snippetstrackerissue'] = 'Tracker issue';
