+++ tests/fixtures/snippets/snippet_usagenote.scss
@@ -2,7 +2,7 @@
  * Snippet Title: Snippet with usage note
  * Scope:         global
  * Goal:          eyecandy
- * Description:   Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+ * Description:   Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
  * Creator:       The Boost Union Team
  * Usage note:    Do this, not that. This snippet is intended to be used in a specific way to achieve the desired effect.
  *
